Create Manifest class and YAML schema
Manifest exists as a class wrapper around the YAML-encoded manifest that
serves as buildlsstsw.sh's interface to ltd-mason.

In particular, the manifest provides:

- parsing of YAML text with ruamel.yaml (for better roundtripping of
  YAML comments)
- accessors into the YAML data

A potential schema for the YAML manifest is provided in
tests/demo_science_pipelines_manifest.yaml.

This can be used for live testing, but it is unfortunately specific to
the current lsstsw build.

For DM-4196.