Design S3 upload API
The ltdmason.s3upload.upload() function is intended to be a simple to
use synchronization function to push new files to S3, and delete
deprecated files from S3. We're using boto3 to interact with S3.

This commit lays out the structure of the upload function, and puts in
stubs for helper function and classes. These need to be implemented.

The ObjectManager, in particular, is used to find and delete files from
S3 that no longer exist in the source.

This commit builds in the idea of a version_slug/ directory to partition
different versions of the docs in the same bucket.

For DM-4196.