Implement filename and directory listing on S3
Implements ObjectManager.list_filenames_in_directory() and
ObjectManager.list_dirnames_in_directory().

These do all the necessary path manipulation and comparison to detect
with paths exist at the root of a given directory in an S3 bucket. This
is necessary to implement since S3 treats buckets as a flat object store
with directories inferred from object names. That said, Boto3 does
return 'objects' that are directories; the only way to know these objects
are directories is through the trailing / in the key name.

For DM-4196.