Validate manifest yaml with a schema
The schema, manifest_schema.yaml in the repo root, is a YAML expression
of JSON schema draft 4. We're able to write down YAML schemas using JSON
schema formats because it's the Python object representations that are
validated, not the text.

See http://spacetelescope.github.io/understanding-json-schema/index.html
for more information on JSON schemas.

Manifest has been updated to run validation on instantiation.

The jsonschema package is a new dependency.

Unit tests now test that validation errors are thrown when required
fields are missing, or when the certain data types are incorrect. This
could be expanded.

For DM-4196.