Embed mock package doc repos for testing
These mock packages, designed to look like LSST Stack packages, can be
used for lightweight end-to-end testing of ltd-mason builds. Since
package repos have always been intended to pre-exist on a local file
system, this is viable for testing. Note that the *product* doc repo is
normally git clone'd during a ltd-mason build. The
github.com/lsst-sqre/mock-doc repo is designed to do this.

For DM-4196.