uploader and s3uploader pass through aws creds
Previously the authentication concept for ltd-mason was to put
credentials into ~/.aws/credentials under a named profile. The name of
the credential profile was the passed through to create session objects.

Now variables for aws key ID, secret and profile name are passed through
allowing for authentication both with a named profile in
~/.aws/crednetials or through (say) environment variables read by the
CLI or tests.

Also solves an issue where the session was correctly used by the
ObjectManager when doing bulk deletes (of a directory on S3).