Test environment variable config for cli
read_aws_credentials() is challenging to test because we not only test
the presence of an environment variable, but also the absence. The
delvar monkeypatch method only removes the monkeypatched var; an actual
environment variable of the same name can still leak through. Instead we
put the logic of omitting variables into a function that replaces
os.getenv.