s3upload sets Content-Type by mimetype
The built-in mimetypes module lets us guess both the Content-Type and
Content-Encoding headers (if applicable) that we need to set for our S3
file uploads.

For testing, the AWS/S3/boto3 API is lacking here. Basically we can't
directly get the HTTP headers that S3 would serve from the S3 API. So
instead we do a HEAD request via HTTP and inspect the headers that way.
Even getting the URL of the object that S3 serves over HTTP can be
tricky.