Add ltd-mason-travis for Travis-based builds
This CLI can be used specifically for travis-based build environments
and uses the specialized TravisManifest and TravisProduct classes.

Only build on a TRAVIS_PULL_REQUEST=false. A regular branch build is the
only way to know the current TRAVIS_BRANCH (and to avoid build
duplication).

Only build on LTD_MASON_BUILD=true. Again, this helps avoid duplication
in a Travis build matrix; only one entry in the matrix has this as `true`.