Run LTD doc build with Travis
This is a proof-of-concept for supporting documentation publication
through Travis to LSST the Docs. I'd expect this to be the main way that
technotes, the Developer Guide and standalone Python package
documentation will be published.

In this Travis, I add an extra build (outside the matrix) where I
specifically enable the ltd-mason-travis upload through the
LTD_MASON_BUILD environment variable. This is needed since otherwise
Travis will have Mason upload a doc build for each entry in the build
matrix!

Also note that the Sphinx build is carried out under Travis instead of
part of LTD Mason. This is done because Travis is already great at
running command line tasks, and also to verify that the docs build on
each platform.