Recommend that doctree be set in Travis builds
The sphinx-build -d flag indicates the directory of the doctree cache
output. If not set, this defaults to the same directory as the html
output, which then gets uploaded to S3. Setting this flag to the normal
location that the Makefile sets it to prevents this.

Includes fixes for ltd-mason's own travis configuration and the Travis
setup documentation.